# docs/swagger.yaml
definitions:
...
  errno.Errno:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  model.History:
    properties:
      name:
        type: string
      user_id:
        type: string
    type: object
...
  search.UserInfo:
    properties:
      avatar:
        type: string
      energy:
        type: integer
      id:
        type: integer
      nickname:
        type: string
    type: object
  search.UserListResponse:
    properties:
      count:
        type: integer
      list:
        items:
          $ref: '#/definitions/search.UserInfo'
        type: array
    type: object
  user.GetUserInfoResponse:
    properties:
      avatar:
        type: string
      energy:
        type: integer
      id:
        type: integer
      nickname:
        type: string
    type: object
...
host: 124.71.184.107
info:
  contact:
    email: 2036479155@qq.com
    name: gongna
  description: 零重力APP
  termsOfService: http://swagger.io/terrms
  title: ZeroGravity
  version: 1.0.0
paths:
...
  /api/v1/search:
    delete:
      consumes:
      - application/json
      description: 删除查询历史
      parameters:
      - description: userId--用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: delete history删除历史记录
        in: body
        name: h
        required: true
        schema:
          $ref: '#/definitions/model.History'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errno.Errno'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errno.Errno'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errno.Errno'
      summary: 删除查询历史
      tags:
      - search
    get:
      consumes:
      - application/json
      description: 获取搜索结果，结果为用户列表或想法列表
      parameters:
      - description: page--限制指定要检索的记录数
        in: query
        name: page
        required: true
        type: integer
      - description: limit--偏移量指定开始返回记录之前要跳过的记录数
        in: query
        name: limit
        required: true
        type: integer
      - description: 搜索目标，0->idea 1->user 默认搜索 idea
        in: query
        name: target
        required: true
        type: integer
      - description: 关键字
        in: query
        name: keyword
        required: true
        type: string
      - description: userId
        in: path
        name: id
        required: true
        type: integer
      - description: 后端给前端的token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":{"username":"kong"}}'
          schema:
            $ref: '#/definitions/search.UserListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errno.Errno'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errno.Errno'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errno.Errno'
      summary: 获取搜索结果
      tags:
      - search
  /api/v1/search/histories:
    get:
      consumes:
      - application/json
      description: 获取历史记录
      parameters:
      - description: userId--用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":{"username":"kong"}}'
          schema:
            items:
              $ref: '#/definitions/model.History'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errno.Errno'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errno.Errno'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errno.Errno'
      summary: 获取历史记录
      tags:
      - search
...
  /api/v1/user/detail/:id:
    get:
      consumes:
      - application/json
      description: 得到用户所有的个人信息
      parameters:
      - description: userId--用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":{"username":"kong"}}'
          schema:
            $ref: '#/definitions/user.GetUserInfoResponse'
      summary: 得到用户信息
      tags:
      - user
schemes:
- http
swagger: "2.0"